requirements.txt
asgiref==3.7.2
diploma-frontend @ file:///mnt/d/coding/works/django_diploma/diploma-frontend/dist/diploma-frontend-0.6.tar.gz#sha256=7ca983d6d44b4293575841cbfabd0c84312ab6676bd0898d2200ea5131203740
Django==4.2
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
